python/requirements.txt
python-dotenv
openai
freeplay
